% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STTileMaker.R
\name{read_data}
\alias{read_data}
\title{Make Tiles of each Spot in a Spatial Transcriptomics Dataset}
\usage{
read_data(directory)
}
\arguments{
\item{directory, }{directoryectory of ST data}
}
\value{
Tiles??
}
\description{
Make Tiles of each Spot in a Spatial Transcriptomics Dataset
}
\details{
This function will make Tiles of each Spot in a Spatial Transcriptomics Dataset when given
the correct input
}
\examples{
# Make Tiles of each Spot in a Spatial Transcriptomics Dataset
directory = here("data/f12hr_140_processed/outs/")
tiles = STTileMaker(directory)
}
