% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STTilemaker_make_tiles.R
\name{STTileMaker_make_tiles}
\alias{STTileMaker_make_tiles}
\title{Make tiles of each spot in a Spatial Transcriptomics Dataset}
\usage{
STTileMaker_make_tiles(data)
}
\arguments{
\item{data, }{object from function STTileMaker_read_data}
}
\value{
H&E tiles from each spot is returned as a list of lists
}
\description{
Make tiles of each spot in a Spatial Transcriptomics Dataset
}
\details{
This function will make Tiles of each Spot in a Spatial Transcriptomics Dataset when given
the correct input
}
\examples{
tiles = STTileMaker_make_tiles(data)
}
