% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STTileMaker_plot_tile.R
\name{STTileMaker_plot_tile}
\alias{STTileMaker_plot_tile}
\title{Plot a tile of a spot in a Spatial Transcriptomics Dataset}
\usage{
STTileMaker_plot_tile(tiles, index = 1)
}
\arguments{
\item{tiles, }{tiles object from function STTileMaker_make_tiles}

\item{index, }{the index of a tile you want to visualize, range is 1 to n number of tiles}
}
\value{
Plot of a single tile
}
\description{
Plot a tile of a spot in a Spatial Transcriptomics Dataset
}
\details{
This function will make plot a tile of a spot in a Spatial Transcriptomics Dataset as a quality control check
}
\examples{
STTileMaker_plot_tile(tiles, 10)
}
