% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STTileMaker_read_data.R
\name{STTileMaker_read_data}
\alias{STTileMaker_read_data}
\title{Reads in the data that contains the image, gene expression, and various other variables}
\usage{
STTileMaker_read_data(directory)
}
\arguments{
\item{directory, }{directory where the ST data is located}
}
\value{
list of lists with:

pos - original positions of spots

posAligned - positions of spots after data alignment

posClean - positions of spots after data cleaning

img - the H&E image in matrix form

geneCounts - matrix of the of the raw data

geneCountsClean - matrix that has been cleaned with standard ST data methods

r - scale for resizing spots

scalefactor - scalefactor from image to cartesian scales
}
\description{
Reads in the data that contains the image, gene expression, and various other variables
}
\details{
This function will reads in the data that contains the image, gene expression, and various other variables
}
\examples{
data = STTileMaker_read_data(here::here("data/f12hr_140_processed/outs/"))
}
